0909/requirements.txt
discord.py
openai
google-generativeai
python-dotenv
requests
notion-client
Pillow
mistralai==0.4.2
aiohttp
google-cloud-aiplatform>=1.49.0
PyPDF2
fastapi
uvicorn